Feature/content permissions

+++ api/src/modules/content/controllers/content.controller.ts
@@ -1,4 +1,4 @@
-import { Controller, Get, Body, Post, Put, Delete, Param, Headers, Query, UseGuards, UnauthorizedException } from '@nestjs/common';
+import { Controller, Get, Body, Post, Put, Delete, Param, Headers, Query, UseGuards, UnauthorizedException, Request, ForbiddenException } from '@nestjs/common';
 import { ApiBasicAuth, ApiTags } from '@nestjs/swagger';
 
 import { Paginated } from '~shared/types';
@@ -6,7 +6,7 @@ import { AuthGuard } from '~shared/guards/auth.guard';
 import { Permissions, AuditLog } from '~shared/decorators';
 
 import { ContentService } from '../services/content.service';
-import { ContentTypeService } from '../services/content-type.service';
+import { PermissionService } from '~shared/services/permission.service';
 
 
 @Controller('/content-types/:contentTypeUuid/content')
@@ -17,56 +17,76 @@ export class ContentController {
 
 	constructor(
 		private contentService: ContentService,
-		private contentTypeService: ContentTypeService,
+		private permissionService: PermissionService,
 	) { }
 
 	@Get()
-	@Permissions('content/read')
-	public findContentTypeEntries(
+	public async findContentTypeEntries(
 		@Query('page') page = 1,
 		@Query('pagesize') pagesize = 20,
 		@Query('search') search: string,
 		@Param('contentTypeUuid') contentTypeUuid: string,
+		@Request() req,
 	): Promise<Paginated<any>> {
+		if (!await this.permissionService.hasPermission(req.user?.uuid || req.headers.authorization, [`content/${contentTypeUuid}/read`])) {
+			throw new ForbiddenException(`Missing permissions: content/${contentTypeUuid}/read`)
+		}
+
 		return this.contentService.findByContentType(contentTypeUuid, page, pagesize, search);
 	}
 
 	@Get('/:entryUuid')
-	@Permissions('content/read')
 	public async one(
 		@Param('contentTypeUuid') contentTypeUuid: string,
 		@Param('entryUuid') entryUuid: string,
+		@Request() req,
 	): Promise<any | undefined> {
+		if (!await this.permissionService.hasPermission(req.user?.uuid || req.headers.authorization, [`content/${contentTypeUuid}/read`])) {
+			throw new ForbiddenException(`Missing permissions: content/${contentTypeUuid}/read`)
+		}
+
 		return this.contentService.findOne(contentTypeUuid, entryUuid);
 	}
 
 	@Post()
-	@Permissions('content/create')
 	public async create(
 		@Param('contentTypeUuid') contentTypeUuid: string,
 		@Body() content: any,
+		@Request() req,
 	): Promise<any> {
+		if (!await this.permissionService.hasPermission(req.user?.uuid || req.headers.authorization, [`content/${contentTypeUuid}/create`])) {
+			throw new ForbiddenException(`Missing permissions: content/${contentTypeUuid}/read`)
+		}
+
 		return this.contentService.create(contentTypeUuid, content);
 	}
 
 	@Put('/:entryUuid')
-	@Permissions('content/update')
 	@AuditLog('content/update')
 	public async update(
 		@Param('contentTypeUuid') contentTypeUuid: string,
 		@Param('entryUuid') uuid: string,
 		@Body() content: any,
+		@Request() req,
 	): Promise<any> {
+		if (!await this.permissionService.hasPermission(req.user?.uuid || req.headers.authorization, [`content/${contentTypeUuid}/update`])) {
+			throw new ForbiddenException(`Missing permissions: content/${contentTypeUuid}/read`)
+		}
+
 		return this.contentService.update(contentTypeUuid, uuid, content);
 	}
 
 	@Delete('/:entryUuid')
-	@Permissions('content/delete')
 	@AuditLog('content/delete')
 	public async delete(
 		@Param('contentTypeUuid') contentTypeUuid: string,
-		@Param('entryUuid') uuid: string
+		@Param('entryUuid') uuid: string,
+		@Request() req,
 	): Promise<void> {
+		if (!await this.permissionService.hasPermission(req.user?.uuid || req.headers.authorization, [`content/${contentTypeUuid}/delete`])) {
+			throw new ForbiddenException(`Missing permissions: content/${contentTypeUuid}/read`)
+		}
+
 		return await this.contentService.delete(contentTypeUuid, uuid);
 	}
